+++ shared/src/lib.rs
@@ -12,3 +12,11 @@ pub mod shutdown;
 pub mod transports;
 #[cfg(feature = "version")]
 pub mod version;
+
+#[inline]
+pub fn mutex_lock<T>(mutex: &std::sync::Mutex<T>) -> std::sync::MutexGuard<'_, T> {
+    match mutex.lock() {
+        Ok(lock) => lock,
+        Err(p_err) => p_err.into_inner(),
+    }
+}